Refactor: mysql field 변경 및 test.txt 삭제
불필요한 txt파일 삭제와
분명한 필드명을 분명하게 테이블 명을 포함시켜 정의함
