+++ src/main/java/home/todayhome/entity/Comment.java
@@ -33,7 +33,7 @@ public class Comment {
     private User user;
 
     @NotNull
-    @Column(name = "content", nullable = false, length = 200)
+    @Column(name = "comment_contents", nullable = false, length = 200)
     private String content;
 
     @NotNull
