+++ src/main/java/home/todayhome/repository/BoardLikeRepository.java
@@ -1,4 +1,4 @@
-package home.todayhome.repository.board;
+package home.todayhome.repository;
 
 import home.todayhome.entity.LikesBoard;
 import org.springframework.data.jpa.repository.JpaRepository;
