--- src/main/java/home/todayhome/repository/board/BoardRepository.java
@@ -1,7 +1,0 @@
-package home.todayhome.repository.board;
-
-import home.todayhome.entity.Board;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface BoardRepository extends JpaRepository<Board, Integer> {
-}
